refactor(Move some setup code into the setup section):
