+++ src/setup/add-entities-to-spot.ts
@@ -1,0 +1,77 @@
+import { BathingspotMeasurement, BathingspotPrediction, Bathingspot } from '../orm/entity';
+import { Connection, getRepository } from 'typeorm';
+
+// const isPrediction = (entity: BathingspotMeasurement|BathingspotPrediction): entity is BathingspotPrediction {
+//   return (<BathingspotPrediction>entity).prediction !== undefined;
+// }
+
+// ███████╗███████╗████████╗██╗   ██╗██████╗
+// ██╔════╝██╔════╝╚══██╔══╝██║   ██║██╔══██╗
+// ███████╗█████╗     ██║   ██║   ██║██████╔╝
+// ╚════██║██╔══╝     ██║   ██║   ██║██╔═══╝
+// ███████║███████╗   ██║   ╚██████╔╝██║
+// ╚══════╝╚══════╝   ╚═╝    ╚═════╝ ╚═╝
+
+export interface IAddEntitiesToSpotOptions {
+  entities: BathingspotMeasurement[] | BathingspotPrediction[];
+  connection: Connection;
+}
+export type AddEntitiesToSpot = (options: IAddEntitiesToSpotOptions) => Promise<void>;
+
+// used in app.ts for setup
+export const addEntitiesToSpot: AddEntitiesToSpot = async (options) => {
+  try {
+    const spotRepo = getRepository(Bathingspot);
+    for (const entity of options.entities) {
+      const fOpts = { where: {} };
+
+      switch (true) {
+        case entity instanceof BathingspotMeasurement:
+          fOpts.where = {
+            detailId: (entity as BathingspotMeasurement).detailId,
+          };
+          break;
+        case entity instanceof BathingspotPrediction:
+
+          fOpts.where = {
+            oldId: (entity as BathingspotPrediction).oldId,
+          };
+          break;
+      }
+
+      const bspot = await spotRepo.findOne(fOpts);
+
+      if (bspot !== undefined) {
+
+        if (entity instanceof BathingspotPrediction) {
+          // console.log('got prediction');
+          if (bspot.predictions === undefined) {
+            bspot.predictions = [entity];
+          } else {
+            bspot.predictions.push(entity);
+          }
+
+        } else if (entity instanceof BathingspotMeasurement) {
+          if (bspot.measurements === undefined) {
+            bspot.measurements = [entity];
+          } else {
+            bspot.measurements.push(entity);
+          }
+        }
+        await options.connection.manager.save(entity);
+        await options.connection.manager.save(bspot);
+      }
+    }
+  } catch (error) {
+    throw error;
+  }
+};
+
+// const addEntitiy = <E>(entity: E, spot: Bathingspot) =>{
+//   let arr = entity instanceof BathingspotPrediction ?
+//   spot.predictions : entity instanceof BathingspotMeasurement ? spot.measurements : new Error('Neither BathingspotMeasurement nor BathingspotPrediction');
+
+//   if(arr === undefined){
+//     arr = [entity] as (enitity instanceof );
+//   }
+// }
